+++ src/app/auth/login/use-login-form.ts
@@ -1,10 +1,17 @@
+import { useRouter } from 'next/navigation';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
+import { useMutation } from '@tanstack/react-query';
+import { toast } from 'sonner';
 
 import { LoginFormSchema } from '@/schemas/auth/login.schema';
 import type { LoginForm } from '@/schemas/auth/login.schema';
 
+import { signIn } from './sign-in.action';
+
 export default function useLoginForm() {
+  const router = useRouter();
+
   const form = useForm<LoginForm>({
     defaultValues: {
       email: '',
@@ -13,8 +20,19 @@ export default function useLoginForm() {
     resolver: zodResolver(LoginFormSchema),
   });
 
+  const loginFormMutation = useMutation({
+    mutationFn: signIn,
+    onSuccess: (result) => {
+      if (result.success) {
+        router.push('/');
+      } else {
+        toast.error('Invalid login credentials');
+      }
+    },
+  });
+
   const onSubmit = async (data: LoginForm) => {
-    console.log(data); // temp
+    loginFormMutation.mutate(data);
   };
 
   return { form, onSubmit };