+++ CHANGELOG.md
@@ -13,10 +13,12 @@ All notable changes to this project will be documented in this file.
 - login form hook using react hook form
 - forgot password form hook using react hook form
 - reset password form hook using react hook form
+- login form sign in action using supabase auth
 
 #### Lib
 
 - supabase nextjs client
+- supabase log event edge function invoke
 
 #### Schemas
 
@@ -36,6 +38,7 @@ All notable changes to this project will be documented in this file.
 - login form link to forgot password to include auth
 - forgot password form link to login to include auth
 - update password form file name to reset password form
+- login form hook to use tanstack query
 
 #### Components